+++ CMakeLists.txt
@@ -1,9 +1,9 @@
 cmake_minimum_required(VERSION 3.7)
-project(F4MPClient)
+project(F4MP_Client)
 
 set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} CMakeModules/)
 
-set(EXECUTABLE F4MPClient)
+set(EXECUTABLE F4MP_Client)
 set(SRC_DIR ./src)
 set(INCLUDE_DIR ./include)
 
@@ -12,6 +12,8 @@ message(${CMAKE_CURRENT_SOURCE_DIR}/resources)
 
 set(THREADS_PREFER_PTHREAD_FLAD ON)
 find_package(Threads REQUIRED)
+
+add_subdirectory (libs/imgui)
 
 include_directories(${EXECUTABLE}
 	${INCLUDE_DIR}
@@ -23,7 +25,6 @@ file(GLOB SourceFiles
 	${SRC_DIR}/Utils/*
 	${SRC_DIR}/ThirdParty/*
 	${SRC_DIR}/Hooks/*
-	${SRC_DIR}/imgui/*
 )
 
 add_subdirectory(Common)
@@ -31,7 +32,7 @@ add_subdirectory(Common)
 add_library(${EXECUTABLE} SHARED ${SourceFiles})
 
 set_target_properties(${EXECUTABLE} PROPERTIES
-	CXX_STANDARD 20
+	CXX_STANDARD 17
 	CXX_EXTENSIONS ON
 )
 
@@ -44,4 +45,5 @@ target_link_libraries(${EXECUTABLE}
 	Common
 	d3d11.lib
 	dxgi.dll
+	f4mp::imgui
 )
