+++ src/DirectXHook.h
@@ -6,10 +6,10 @@
 //hooks
 #include "Hooks/Direct3D11.h"
 #include "Hooks/DXGI.h"
-//imgui
-#include "imgui/imgui.h"
-#include "imgui/imgui_impl_dx11.h"
-#include "imgui/imgui_impl_win32.h"
+
+#include <imgui.h>
+#include <imgui_impl_dx11.h>
+#include <imgui_impl_win32.h>
 
 using namespace Exceptions::Core::Exceptions;
 using namespace Hooks;
